requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Deprecated==1.2.13
Django==4.1
django-oauth-toolkit==2.1.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-spectacular==0.24.1
drf-spectacular-sidecar==2022.9.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
jwcrypto==1.4.2
MarkupSafe==2.1.1
oauthlib==3.2.1
openapi-codec==1.3.2
pycparser==2.21
Pygments==2.13.0
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
